+++ src/main/java/nextstep/config/GlobalExceptionHandler.java
@@ -1,4 +1,4 @@
-package nextstep;
+package nextstep.config;
 
 import auth.exception.AuthenticationException;
 import nextstep.exception.RoomEscapeException;
