+++ src/main/java/nextstep/member/UserAuthenticatorImpl.java
@@ -4,12 +4,10 @@ import auth.UserAuthenticator;
 import auth.entity.UserDetails;
 import lombok.RequiredArgsConstructor;
 import nextstep.exception.MemberException;
-import org.springframework.stereotype.Component;
 
 import java.util.Optional;
 
 @RequiredArgsConstructor
-@Component
 public class UserAuthenticatorImpl implements UserAuthenticator {
     private final MemberService memberService;
